+++ .prettierrc
@@ -3,5 +3,5 @@
   "singleQuote": true,
   "tabWidth": 2,
   "printWidth": 100,
-  "trailingComma": "es2016"
+  "trailingComma": "es5"
 }
