+++ 03_trip-trip/day22/package.json
@@ -13,6 +13,7 @@
     "graphql": "^16.11.0",
     "next": "14.2.32",
     "react": "^18",
+    "react-daum-postcode": "^3.2.0",
     "react-dom": "^18",
     "react-router": "^7.8.2"
   },
