+++ 03_trip-trip/day22/src/features/boards/ui/BoardForm.tsx
@@ -2,9 +2,10 @@
 import useForm from '@/features/boards/hooks/useBoardForm'
 import { initialPostValue } from '@/features/boards/model/types'
 import validate from '@/features/boards/model/validation'
-import CustomButton from '@/shared/ui/custom-button/CustomButton'
-import UploadImagesWithLabel from '@/shared/ui/upload-image/UploadImagesWithLabel'
+import CustomButton from '@/shared/ui/CustomButton/CustomButton'
+import UploadImagesWithLabel from '@/shared/ui/UploadImage/UploadImagesWithLabel'
 import styles from './BoardForm.module.css'
+import AddressForm from '@/shared/ui/AddressForm/AddressForm'
 
 export default function BoardForm() {
   const {
@@ -14,6 +15,7 @@ export default function BoardForm() {
     handleChange,
     handleSubmit,
   } = useForm({ initialValues: initialPostValue, validate })
+
   return (
     <form className={styles['post-form']} onSubmit={handleSubmit}>
       <div className={styles['post-form-col']}>
@@ -22,7 +24,7 @@ export default function BoardForm() {
           <div>
             <label>작성자</label>
             <span>*</span>
-          </div> 
+          </div>
 
           <input
             type="text"
@@ -87,22 +89,7 @@ export default function BoardForm() {
       </div>
 
       {/* 주소 */}
-      <div className={styles['post-form-addr']}>
-        <div>
-          <label>주소</label>
-        </div>
-        <div className={styles['post-form-addr-active']}>
-          <input
-            className={styles['post-form-addr-zipcode']}
-            placeholder="01234"
-            name="addr.zipcode"
-            onChange={handleChange}
-          />
-          <CustomButton type={'button'} content={'우편번호 검색'} color={'default'} />
-        </div>
-        <input placeholder="주소를 입력해 주세요." name="addr.addr1" onChange={handleChange}/>
-        <input placeholder="상세주소" name="addr.addr2" onChange={handleChange}/>
-      </div>
+      <AddressForm onChange={handleChange} value={post.addr} />
       <hr />
       {/* 유튜브 링크 */}
       <div className={styles['post-form-input']}>
