+++ 03_trip-trip/day22/src/shared/ui/AddressForm/AddressForm.module.css
@@ -1,0 +1,13 @@
+/* AddrWithLabel */
+.post-form-addr {
+  display: flex;
+  flex-direction: column;
+  gap: 0.8rem;
+}
+.post-form-addr-active {
+  display: flex;
+  gap: 0.8rem;
+}
+.post-form-addr-zipcode {
+  width: 8.2rem;
+}
