+++ 03_trip-trip/day22/src/shared/ui/AddressForm/types.ts
@@ -1,0 +1,5 @@
+export interface PostAddr {
+  zipcode?: string
+  addr1?: string
+  addr2?: string
+}
