+++ 03_trip-trip/day22/src/shared/ui/CustomButton/CustomButton.types.ts
