+++ 03_trip-trip/day22/src/shared/ui/CustomButton/lib/styles.ts
