+++ 03_trip-trip/day22/src/shared/ui/UploadImage/UploadImage.module.css
