+++ 03_trip-trip/day22/src/shared/ui/UploadImage/UploadImagesWithLabel.tsx
